Finished the newCustomTime() fn.
This function records the time.now, then calls a function to parse the
time into an array while singling out the minute portion and rounding it
down to the closest 5 minute increment. An example would be 47 mins would round
down to 45 minutes and 22 mins would round down to 20 mins.

Also the if statement comparison between time_now_rounded & old_time may be
somewhat terse so I will talk it through now.
The statement is '(time_now_rounded - old_time).empty? == true'
So the overall comparison is a boolean == true
The (time_now_rounded - old_time).empty? is subtracting two arrays and if all
the elements values are exactly the same then the array will equal an empty
array '[]'. So the .empty? method will return true if it is called on an empty
array.