The Streams controller.
This controller has two functions. The first calls the twitch api and
retrieves data that is parsed into an array of 20 hashes that represent
the top streamers currently. The second function calls the first and then
attempts to place some info in the top_streams_static db table if an entry
is not already there. Then it places all 20 entries in the top_streams db
table.