Add controller to handle retrieving games.
Like the topStreams.rb controller this controller has two functions
in which one makes an api call to Twitch for data and the second function
parses the data and generates a query for the model. Next the second function
is checking to see if the data has been entered into our top_games_static db
table and if it has not it will enter it there. And lastly it will enter the top
20 games into the top_games db table.