The main db model for top_games table.
This file contains two functions that work together to build an entry
to the top_games table. The first function is assembling a query and parsing
the data to the model into a multidimensional array. Both these items are
returned in a hash. The second function executes the command with the returned
data. There is a definate oppurtunity to pull the parsing portion of the first
function into its own function.