Controller for handling my parsing of time.now().
The summary of this portion is that I want this code to run every five
minutes based on a cron cycle. So I have a check to see if the last
entry in my db is the same as this current time.now() and if it is the
same then we do now execute the rest of our code but just close the
connection to then db. Also we create a new connection to the db every
time we start this microservice and close it at the end of our execution
cycle.